Ticket 50077 - Fix compiler warnings in automember rebuild task
https://pagure.io/389-ds-base/issue/50077

Reviewed by: mreynolds(one line commit rule)

(cherry picked from commit eae696f1e0cf477f96966859feab9eacd9a0be82)

+++ ldap/servers/plugins/automember/automember.c
@@ -2455,8 +2455,8 @@ out:
         slapi_task_log_notice(task, "Automember rebuild task aborted.  Error (%d)", result);
         slapi_task_log_status(task, "Automember rebuild task aborted.  Error (%d)", result);
     } else {
-        slapi_task_log_notice(task, "Automember rebuild task finished. Processed (%d) entries.", i);
-        slapi_task_log_status(task, "Automember rebuild task finished. Processed (%d) entries.", i);
+        slapi_task_log_notice(task, "Automember rebuild task finished. Processed (%d) entries.", (int32_t)i);
+        slapi_task_log_status(task, "Automember rebuild task finished. Processed (%d) entries.", (int32_t)i);
     }
     slapi_task_inc_progress(task);
     slapi_task_finish(task, result);
